Add tab and tabs components. Add mobile design for tabs mobile navbar
