+++ src/components/Home/AdvancedSearch.jsx
@@ -30,7 +30,7 @@ const Select = ({ value, onChange, options, keys = false }) => (
   </select>
 );
 
-const AdvancedSearch = () => {
+const AdvancedSearch = ({showFilters = true}) => {
   const {
     forms: { search },
     games,
@@ -59,33 +59,63 @@ const AdvancedSearch = () => {
 
   return (
     <div className="browse-listings flex-column flex-grow">
-      <div id="navbarBasicExample" className="flex-row flex-grow search-filters">
-        <div className="navbar-start flex-row flex-grow">
-          <div className="search-item flex-row align-center">
-            <label htmlFor="search-sort-by">Sort By:</label>
-            <Select 
-              name="sort_by"
-              id="search-sort-by"
-              value={search.sort_by}
-              onChange={handleChange}  
-              options={sortByOptions}
-              keys={true}
-            />
+      <div id="navbarBasicExample" className="flex-row flex-grow justify-flex-end search-filters">
+
+       
+       {showFilters &&
+        <div className="navbar-start flex-column flex-grow">
+
+          <div className="flex-row flex-grow">
+            <div className="search-item flex-row align-center">
+              <label htmlFor="search-sort-by">Sort By:</label>
+              <Select 
+                name="sort_by"
+                id="search-sort-by"
+                value={search.sort_by}
+                onChange={handleChange}  
+                options={sortByOptions}
+                keys={true}
+              />
+            </div>
+
+            <div className="search-item flex-row align-center">
+              <label htmlFor="search-platform">Platform:</label>
+              <Select 
+                name="platform"
+                id="search-platform"
+                value={search.platform}
+                onChange={handleChange} 
+                options={platformOptions}
+                keys={true}
+              />
+            </div>
           </div>
 
-          <div className="search-item flex-row align-center">
-            <label htmlFor="search-platform">Platform:</label>
-            <Select 
-              name="platform"
-              id="search-platform"
-              value={search.platform}
-              onChange={handleChange} 
-              options={platformOptions}
-              keys={true}
-            />
+          <div className="flex-row more-filters align-center">
+            <div className="search-item">
+              <label htmlFor="search-genre">Genre:</label>
+              <Select 
+                name="genre"
+                id="search-genre"
+                value={search.genre}
+                onChange={handleChange}
+                options={genreOptions}
+                keys={false}
+              />
+            </div>
+
+            <div className="search-item flex-row align-center">
+              <label htmlFor="price-range">Price Range:</label>
+              <RangeSlider
+                range={search.price.defaultRange()}
+                values={search.price.values()}
+                setValues={setPriceValues}
+              />
+            </div>
           </div>
 
         </div>
+      }
 
   
 
@@ -94,31 +124,11 @@ const AdvancedSearch = () => {
               <SearchInput onChange={handleChange} value={search.query}  name="query"/>
             </form>
         </div>
-      </div>
 
-      <div className="flex-row justify-between more-filters">
-        <div className="search-item">
-          <label htmlFor="search-genre">Genre:</label>
-          <Select 
-            name="genre"
-            id="search-genre"
-            value={search.genre}
-            onChange={handleChange}
-            options={genreOptions}
-            keys={false}
-          />
-        </div>
-
-        <div className="search-item flex-row align-center">
-          <label htmlFor="price-range">Price Range:</label>
-          <RangeSlider
-            range={search.price.defaultRange()}
-            values={search.price.values()}
-            setValues={setPriceValues}
-          />
-        </div>
 
       </div>
+
+      
     </div>
   );
 };
