+++ src/components/Home/DropDown/DropDown.scss
@@ -2,7 +2,7 @@
   position: absolute;
   top: 0;
   left: 0;
-  width: 100vw;
+  width: 100%;
   
   box-shadow: 0px 1px 1px #FFFFFF26;
 
@@ -46,7 +46,7 @@
     z-index: 999;
     width: 100vw;
     height: 100vh;
-    background-color: rgba(18, 18, 18, 0.55);
+    background-color: rgba(18, 18, 18, 0.90);
   }
 
   .tabs-mobile {
