+++ src/components/SearchBar/SearchBar.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
 
-import SearchInput from '../Form/SearchInput/SearchInput';
+import AdvancedSearch from '../Home/AdvancedSearch';
 
-const SearchBar = ({ children, show = true }) => (
-  <div className="flex-row justify-between justify-flex-start flex-grow align-center top-search-bar">
+const SearchBar = ({ children, onSubmit, show = true }) => (
+  <div className="flex-row justify-between flex-grow align-center top-search-bar">
 
     {children}
 
     { 
       show &&
       <div className="flex-row flex-grow justify-flex-end">
-        <form>
-          <SearchInput />
-        </form>
+
+          <AdvancedSearch showFilters={false}/>
+
       </div>
     }
